Add random sized rooms, don't print dead doors
Created the room randomizing function.  Created room type "hall" which
is not counted toward the "max number of rooms" when the dungeon is
created.  Flagged doors with a "print"/"don't print" and added a check
for that in the door print function.  Added a "this is the door on the
other side" variable to the door class, used to print neither door if it
leads to nowhere.